+++ reportbro/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.11.2"
+__version__ = "0.12.1"
 
 from .reportbro import Report, ReportBroError